requirements.txt
# Core data processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Model monitoring (Evidently AI)
# Using compatible version with Airflow 2.7.3 / Pydantic
evidently==0.4.33
pydantic==2.5.3

# Configuration and utilities
pyyaml==6.0.1
python-dateutil==2.8.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Serialization
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.0
pylint==3.0.3

# Airflow (will be installed in Docker)
# apache-airflow==2.7.0
